Refactor Jetsnack theme to use a unified `JetsnackTheme` class and one `CompositionLocal`.
- Deleted `CompositionLocals.kt` which used reflection to access internal Material 3 locals.
- Introduced `JetsnackTheme` class to encapsulate `JetsnackColors`, `Typography`, `Shapes`, and `AppStyles`.
- Replaced individual `LocalJetsnackColors`, `LocalAppStyles`, and Material 3 `LocalTypography`/`LocalShapes` lookups with `JetsnackTheme.colors`, `JetsnackTheme.typography`, etc.
- Added a `StyleScope.currentJetsnackTheme` helper to provide easy access to the theme within `Style` blocks.
- Moved `JetsnackColors` and color palette definitions to `Color.kt`.
- Updated `debugColors` to include newer Material 3 color roles.
