+++ Jetsnack/app/src/main/java/com/example/jetsnack/ui/theme/Styles.kt
@@ -6,63 +6,60 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.style.ExperimentalFoundationStyleApi
 import androidx.compose.foundation.style.Style
+import androidx.compose.foundation.style.StyleScope
 import androidx.compose.foundation.style.disabled
 import androidx.compose.foundation.style.fillWidth
-import androidx.compose.foundation.style.hovered
 import androidx.compose.foundation.style.pressed
 import androidx.compose.foundation.style.selected
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.LocalShapes
 import androidx.compose.material3.LocalTextStyle
-import androidx.compose.material3.LocalTypography
-import androidx.compose.material3.Typography
 import androidx.compose.runtime.Immutable
-import androidx.compose.runtime.ProvidableCompositionLocal
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.TileMode
 import androidx.compose.ui.unit.dp
 import com.example.jetsnack.ui.components.jetsnackTextStyle
+import com.example.jetsnack.ui.theme.JetsnackTheme.Companion.LocalJetsnackTheme
 
 @Immutable
 data class AppStyles(
     val buttonStyle : Style = Style {
         shape( RoundedCornerShape(percent = 50))
-        background(Brush.linearGradient(LocalJetsnackColors.currentValue.interactivePrimary))
-        contentColor(LocalJetsnackColors.currentValue.textInteractive)
+        background(Brush.linearGradient(currentJetsnackTheme.colors.interactivePrimary))
+        contentColor(currentJetsnackTheme.colors.textInteractive)
         contentPadding( ButtonDefaults.ContentPadding.calculateTopPadding()) // todo file issue to support padding values
-        jetsnackTextStyle(LocalTypography.currentValue.labelLarge)
+        jetsnackTextStyle(currentJetsnackTheme.typography.labelLarge)
         disabled {
             animate {
-                background(Brush.linearGradient(LocalJetsnackColors.currentValue.interactiveSecondary))
-                contentColor(LocalJetsnackColors.currentValue.textHelp)
+                background(Brush.linearGradient(currentJetsnackTheme.colors.interactiveSecondary))
+                contentColor(currentJetsnackTheme.colors.textHelp)
             }
         }
     },
     val cardStyle: Style = Style {
-        shape(LocalShapes.currentValue.medium)
-        background(LocalJetsnackColors.currentValue.uiBackground)
-        contentColor(LocalJetsnackColors.currentValue.textPrimary)
+        shape(currentJetsnackTheme.shapes.medium)
+        background(currentJetsnackTheme.colors.uiBackground)
+        contentColor(currentJetsnackTheme.colors.textPrimary)
         /*
         todo elevation
         elevation: Dp = 4.dp,*/
     },
     val dividerStyle: Style = Style {
-        background(LocalJetsnackColors.currentValue.uiBorder.copy(alpha = 0.12f))
+        background(currentJetsnackTheme.colors.uiBorder.copy(alpha = 0.12f))
         height(1.dp)
         fillWidth()
     },
     val gradientIconButtonStyle: Style = Style {
         shape(CircleShape)
         clip(true)
-        border(2.dp, Brush.linearGradient(LocalJetsnackColors.currentValue.interactiveSecondary))
-        background(LocalJetsnackColors.currentValue.uiBackground)
+        border(2.dp, Brush.linearGradient(currentJetsnackTheme.colors.interactiveSecondary))
+        background(currentJetsnackTheme.colors.uiBackground)
         pressed {
             animate {
                 background(Brush.horizontalGradient(
                     // this was a parameter input into the function? might want to make helper function for it
-                    colors = LocalJetsnackColors.currentValue.interactiveSecondary,
+                    colors = currentJetsnackTheme.colors.interactiveSecondary,
                     startX = 0f,
                     endX = 200f,
                     tileMode = TileMode.Mirror,
@@ -71,14 +68,14 @@ data class AppStyles(
         }
     },
     val filterChipStyle: Style = Style {
-        shape(LocalShapes.currentValue.small)
-        background(LocalJetsnackColors.currentValue.uiBackground)
-        contentColor(LocalJetsnackColors.currentValue.textSecondary)
-        border(1.dp, Brush.linearGradient(LocalJetsnackColors.currentValue.interactiveSecondary))
+        shape(currentJetsnackTheme.shapes.small)
+        background(currentJetsnackTheme.colors.uiBackground)
+        contentColor(currentJetsnackTheme.colors.textSecondary)
+        border(1.dp, Brush.linearGradient(currentJetsnackTheme.colors.interactiveSecondary))
         // todo elevation = 2.dp,
         selected {
             animate {
-                background(LocalJetsnackColors.currentValue.brandSecondary)
+                background(currentJetsnackTheme.colors.brandSecondary)
                 contentColor(Color.Black)
                 border(1.dp, Color.Transparent)
             }
@@ -89,8 +86,10 @@ data class AppStyles(
     },
     val surfaceStyle: Style = Style {
         shape(RectangleShape)
-        background(LocalJetsnackColors.currentValue.uiBackground)
-        contentColor(LocalJetsnackColors.currentValue.textSecondary)
+        background(currentJetsnackTheme.colors.uiBackground)
+        contentColor(currentJetsnackTheme.colors.textSecondary)
         clip(true)
     }
 )
+val StyleScope.currentJetsnackTheme: JetsnackTheme
+    get() = LocalJetsnackTheme.currentValue
